+++ rs/embedders/src/wasm_executor.rs
@@ -777,13 +777,20 @@ pub fn process(
                     trap_code,
                     backtrace,
                 } => match backtrace {
-                    Some(bt) => Some(format!("[TRAP]: {}\nCanister Backtrace: {}", trap_code, bt)),
+                    Some(bt) => Some(format!("[TRAP]: {}\n{}", trap_code, bt)),
                     None => Some(format!("[TRAP]: {}", trap_code)),
                 },
-                HypervisorError::CalledTrap(text) if text.is_empty() => {
-                    Some("[TRAP]: (no message)".to_string())
+                HypervisorError::CalledTrap { message, backtrace } => {
+                    let message = if message.is_empty() {
+                        "(no message)"
+                    } else {
+                        &message
+                    };
+                    match backtrace {
+                        Some(bt) => Some(format!("[TRAP]: {}\n{}", message, bt)),
+                        None => Some(format!("[TRAP]: {}", message)),
+                    }
                 }
-                HypervisorError::CalledTrap(text) => Some(format!("[TRAP]: {}", text)),
                 _ => None,
             } {
                 canister_log.add_record(timestamp_nanos, log_message.into_bytes());