+++ rs/rosetta-api/tests/system_tests/common/system_test_environment.rs
@@ -140,17 +140,11 @@ impl RosettaTestingEnvironment {
             .current_block_identifier
             .index;
 
-        self.rosetta_context.kill();
+        self.rosetta_context.kill_rosetta_process();
 
         let rosetta_bin = path_from_env("ROSETTA_BIN_PATH");
-        let rosetta_state_directory =
-            TempDir::new().expect("failed to create a temporary directory");
-        self.rosetta_context = start_rosetta(
-            &rosetta_bin,
-            Some(rosetta_state_directory.path().to_owned()),
-            options,
-        )
-        .await;
+        self.rosetta_context =
+            start_rosetta(&rosetta_bin, self.rosetta_context.state_directory, options).await;
 
         self.rosetta_client =
             RosettaClient::from_str_url(&format!("http://localhost:{}", self.rosetta_context.port))
@@ -171,6 +165,7 @@ pub struct RosettaTestingEnvironmentBuilder {
     pub minting_account: Option<Account>,
     pub initial_balances: Option<HashMap<AccountIdentifier, icp_ledger::Tokens>>,
     pub governance_canister: bool,
+    pub persistent_storage: bool,
 }
 
 impl RosettaTestingEnvironmentBuilder {
@@ -180,6 +175,7 @@ impl RosettaTestingEnvironmentBuilder {
             minting_account: None,
             initial_balances: None,
             governance_canister: false,
+            persistent_storage: false,
         }
     }
 
@@ -206,6 +202,11 @@ impl RosettaTestingEnvironmentBuilder {
 
     pub fn with_governance_canister(mut self) -> Self {
         self.governance_canister = true;
+        self
+    }
+
+    pub fn with_persistent_storage(mut self, enable_persistent_storage: bool) -> Self {
+        self.persistent_storage = enable_persistent_storage;
         self
     }
 
@@ -347,10 +348,16 @@ impl RosettaTestingEnvironmentBuilder {
         let rosetta_bin = path_from_env("ROSETTA_BIN_PATH");
         let rosetta_state_directory =
             TempDir::new().expect("failed to create a temporary directory");
+
+        let mut rosetta_options_builder = RosettaOptionsBuilder::new(replica_url.to_string());
+
+        if self.persistent_storage {
+            rosetta_options_builder = rosetta_options_builder.with_persistent_storage();
+        }
         let rosetta_context = start_rosetta(
             &rosetta_bin,
-            Some(rosetta_state_directory.path().to_owned()),
-            RosettaOptionsBuilder::new(replica_url.to_string()).build(),
+            rosetta_state_directory,
+            rosetta_options_builder.build(),
         )
         .await;